requirements.txt
# Core dependencies
pydantic>=2.0.0
xmltodict>=0.13.0
python-dateutil>=2.8.2

# For cryptographic operations (ZATCA signatures)
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
